Synchronize all access to a shared database connection
In system tests, a single database connection is shared among
all the server threads. Since connections have per-instance
mutable state, this leads to race conditions, as in #36763.

This patch changes the connection pool to wrap shared
connections so that all access to them is synchronized on
the connection's monitor lock.